update the autosde spec to use v1 autosde gem cassette and create v2 cassettes
